property-ai-masterpiece/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
bcrypt
python-jose[cryptography]
pillow
requests==2.31.0
replicate==0.23.0
torch
torchvision
transformers
sentence-transformers
pinecone-client
python-dotenv
opencv-python
ultralytics
streamlit==1.41.0
plotly==5.19.0
datasets
tqdm
scipy
accelerate
numpy
